requirements.txt
# Core web framework and dependencies
Flask==3.0.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3

# LLM model handling
llama-cpp-python==0.2.23
numpy==1.26.2

# Environment and configuration
python-dotenv==1.0.0
typing-extensions==4.9.0

# Testing and development tools
pytest==7.4.3
pytest-cov==4.1.0

# Logging and monitoring
colorama==0.4.6